@@ -80,17 +80,19 @@
     "generate-classes-from-schema": "xsd2ts --force -i schema/XMLSchema.xsd -o shared/generated"
   },
   "devDependencies": {
+    "@eslint/js": "^10.0.1",
     "@types/jest": "^30.0.0",
     "@types/node": "^25.2.3",
     "@types/vscode": "^1.109.0",
-    "@typescript-eslint/eslint-plugin": "^8.54.0",
+    "@typescript-eslint/eslint-plugin": "^8.56.0",
     "@typescript-eslint/parser": "^8.48.1",
     "@vscode/vsce": "^3.7.1",
     "concurrently": "^9.2.1",
     "copy-webpack-plugin": "^13.0.1",
     "css-loader": "^7.1.4",
-    "eslint": "^9.39.2",
-    "eslint-plugin-jest": "^29.12.2",
+    "eslint": "^10.0.0",
+    "eslint-plugin-jest": "^29.15.0",
+    "globals": "^17.3.0",
     "jest": "^30.2.0",
     "jest-environment-jsdom": "^30.2.0",
     "jest-junit": "^16.0.0",
